third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/heap-snapshot-summary-show-ranges-expected.txt

Tests showing several node ranges in the Summary view of detailed heap snapshot.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
Row found: true

Step 3
Retrieved ranges: [{"from":0,"to":10}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] Show 10 beforeShow all 40Show 10 after

Step 4
Retrieved ranges: [{"from":0,"to":10},{"from":30,"to":40}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] Show 10 beforeShow all 20Show 10 after
[11] A 133  3241  241  
[12] A 17  3231  231  
[13] A 145  3231  231  
[14] A 177  3211  211  
[15] A 197  3211  211  
[16] A 189  3181  181  
[17] A 25  3161  161  
[18] A 109  3151  151  
[19] A 201  3151  151  
[20] A 193  3131  131  
[21] Show 10 beforeShow all 10Show 10 after

Step 5
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":30,"to":40}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] Show 10 beforeShow all 10Show 10 after
[11] A 129  3342  342  
[12] A 81  3312  312  
[13] A 157  3302  302  
[14] A 137  3291  291  
[15] A 169  3291  291  
[16] Show 10 beforeShow all 5Show 10 after
[17] A 133  3241  241  
[18] A 17  3231  231  
[19] A 145  3231  231  
[20] A 177  3211  211  
[21] A 197  3211  211  
[22] A 189  3181  181  
[23] A 25  3161  161  
[24] A 109  3151  151  
[25] A 201  3151  151  
[26] A 193  3131  131  
[27] Show 10 beforeShow all 10Show 10 after

Step 6
Retrieved ranges: [{"from":0,"to":10},{"from":20,"to":25},{"from":28,"to":40}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] Show 10 beforeShow all 10Show 10 after
[11] A 129  3342  342  
[12] A 81  3312  312  
[13] A 157  3302  302  
[14] A 137  3291  291  
[15] A 169  3291  291  
[16] Show 10 beforeShow all 3Show 10 after
[17] A 45  3251  251  
[18] A 33  3241  241  
[19] A 133  3241  241  
[20] A 17  3231  231  
[21] A 145  3231  231  
[22] A 177  3211  211  
[23] A 197  3211  211  
[24] A 189  3181  181  
[25] A 25  3161  161  
[26] A 109  3151  151  
[27] A 201  3151  151  
[28] A 193  3131  131  
[29] Show 10 beforeShow all 10Show 10 after

Step 7
Retrieved ranges: [{"from":0,"to":10},{"from":18,"to":26},{"from":28,"to":40}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] Show 10 beforeShow all 8Show 10 after
[11] A 125  3352  352  
[12] A 117  3342  342  
[13] A 129  3342  342  
[14] A 81  3312  312  
[15] A 157  3302  302  
[16] A 137  3291  291  
[17] A 169  3291  291  
[18] A 205  3281  281  
[19] Show 10 beforeShow all 2Show 10 after
[20] A 45  3251  251  
[21] A 33  3241  241  
[22] A 133  3241  241  
[23] A 17  3231  231  
[24] A 145  3231  231  
[25] A 177  3211  211  
[26] A 197  3211  211  
[27] A 189  3181  181  
[28] A 25  3161  161  
[29] A 109  3151  151  
[30] A 201  3151  151  
[31] A 193  3131  131  
[32] Show 10 beforeShow all 10Show 10 after

Step 8
Retrieved ranges: [{"from":0,"to":10},{"from":15,"to":45}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] Show 10 beforeShow all 5Show 10 after
[11] A 65  3382  382  
[12] A 113  3382  382  
[13] A 153  3362  362  
[14] A 125  3352  352  
[15] A 117  3342  342  
[16] A 129  3342  342  
[17] A 81  3312  312  
[18] A 157  3302  302  
[19] A 137  3291  291  
[20] A 169  3291  291  
[21] A 205  3281  281  
[22] A 37  3271  271  
[23] A 49  3261  261  
[24] A 45  3251  251  
[25] A 33  3241  241  
[26] A 133  3241  241  
[27] A 17  3231  231  
[28] A 145  3231  231  
[29] A 177  3211  211  
[30] A 197  3211  211  
[31] A 189  3181  181  
[32] A 25  3161  161  
[33] A 109  3151  151  
[34] A 201  3151  151  
[35] A 193  3131  131  
[36] A 149  3121  121  
[37] A 69  3111  111  
[38] A 29  390  90  
[39] A 121  380  80  
[40] A 165  380  80  
[41] Show 10 beforeShow all 5Show 10 after

Step 9
Retrieved ranges: [{"from":0,"to":45}]
[0] A 77  3492  492  
[1] A 21  3482  482  
[2] A 61  3482  482  
[3] A 85  3472  472  
[4] A 181  3462  462  
[5] A 41  3452  452  
[6] A 141  3452  452  
[7] A 185  3442  442  
[8] A 57  3432  432  
[9] A 53  3412  412  
[10] A 89  3402  402  
[11] A 97  3402  402  
[12] A 73  3392  392  
[13] A 101  3392  392  
[14] A 13  3382  382  
[15] A 65  3382  382  
[16] A 113  3382  382  
[17] A 153  3362  362  
[18] A 125  3352  352  
[19] A 117  3342  342  
[20] A 129  3342  342  
[21] A 81  3312  312  
[22] A 157  3302  302  
[23] A 137  3291  291  
[24] A 169  3291  291  
[25] A 205  3281  281  
[26] A 37  3271  271  
[27] A 49  3261  261  
[28] A 45  3251  251  
[29] A 33  3241  241  
[30] A 133  3241  241  
[31] A 17  3231  231  
[32] A 145  3231  231  
[33] A 177  3211  211  
[34] A 197  3211  211  
[35] A 189  3181  181  
[36] A 25  3161  161  
[37] A 109  3151  151  
[38] A 201  3151  151  
[39] A 193  3131  131  
[40] A 149  3121  121  
[41] A 69  3111  111  
[42] A 29  390  90  
[43] A 121  380  80  
[44] A 165  380  80  
[45] Show 10 beforeShow all 5Show 10 after

Profiler was disabled.
